Refactor FormContainer & actions
