+++ src/actions/auth/auth.ts
@@ -1,64 +1,13 @@
 "use server"
 
-import { connectDB } from "@/db"
-import { encrypt, verifyPassword } from "@/utils/hash"
-import { cookies } from "next/headers"
-import jwt from "jsonwebtoken"
-import { Resend } from "resend"
-import { ObjectId } from "mongodb"
-import { EmailTemplate } from "@/components/auth/AuthLayout/FormContainer/EmailTemplate"
-
-type User = {
-  _id?: ObjectId,
-  name: string,
-  email: string,
-  password: string,
-  createDate: Date,
-  loggedIn: Date,
-  loggedOut: null,
-}
-
-export type SignUpFormData = {
-  name: string,
-  email: string,
-  password: string,
-  rePassword: string
-}
-
-const connection = connectDB()
-
-const createResponse = ({ success, message, data }: { success: boolean, message: string, data?: object }) => {
-  const status = success ? 200 : 500
-  return { success, message, data, status }
-};
-
-const fetchUser = async (email: string) => {
-  const collection = (await connection).collection("user")
-  return { user: await collection.findOne({ email }), collection }
-}
-
-const secretKey = process.env.JWT_SECRET_KEY
-
-if(!secretKey) throw new Error("JWT Secret Key is not valid")
-
-const setCookie = (email: string, status: boolean = false) => {
-  const maxAge = status ? 15 * 24 * 60 * 60 : 24 * 60 * 60
-  const token = jwt.sign({ email }, secretKey);
-  cookies().set("authToken", token, {secure: true, httpOnly: true, maxAge});
-}
-
-const getSerializedUserInfo = (data: object, user: User) => {
-  const serializedUserInfo = {
-    ...data,
-    createDate: user.createDate.toISOString(),
-    loggedIn: user.loggedIn.toISOString(),
-  };
-  return serializedUserInfo
-}
+import { encrypt, verifyPassword } from "@/utils/hash";
+import { UserType } from "@/types/User";
+import { SignUpFormDataType } from "@/types/Form";
+import { createResponse, fetchUser, getSerializedUserInfo, setCookie } from "./common";
 
 export const signup = async (formData: FormData) => {
   try {
-    const form = Object.fromEntries(formData.entries()) as SignUpFormData;
+    const form = Object.fromEntries(formData.entries()) as SignUpFormDataType;
     const { rePassword, ...userDetails } = form;
 
     const { name, email, password } = userDetails;
@@ -69,7 +18,7 @@ export const signup = async (formData: FormData) => {
     // Hash the password
     const hashedPassword: string = await encrypt(password);
 
-    const userInfo: User = {
+    const userInfo: UserType = {
       name: name,
       email: email,
       password: hashedPassword,
@@ -120,7 +69,7 @@ export const signin = async (password: string, email: string, status: boolean) =
   
     // else generate a token, set that in cookie and return true along with data 
     else {
-      const typedUser = user as User
+      const typedUser = user as UserType
       const result = await collection.updateOne({ email }, { $set: { loggedIn: new Date() } })
       if (!result.acknowledged)
         return createResponse({ success: false, message: "Not Signed in", data: {} })
@@ -140,41 +89,12 @@ export const signin = async (password: string, email: string, status: boolean) =
   }
 }
 
-export const checkUser = async (email: string) => {
-  const { user, collection } = await fetchUser(email)
-  if(user) return true
-  return false
-}
-
-export const sendEmail = async(email: string) => {
-
-  const resend = new Resend(process.env.RESEND_API_KEY);
-  try{
-    const otp = generateOTP()
-    const {data, error} = await resend.emails.send({
-      from: 'Amazon-Clone.com <onboarding@resend.dev>',
-      to: email,
-      subject: 'Amazon password assistance',
-      react: EmailTemplate({ otp }) as React.ReactElement,
-    });
-    if(data) return {data, otp}
-    return {data, error, otp}
-  }
-  catch(error){
-    return error
-  }
-}
-
-const generateOTP = () => {
-  return Math.floor(100000 + Math.random() * 900000)
-}
-
 export const changePassword = async(email: string, password: string) => {
   const {user, collection} = await fetchUser(email)
 
   if(!user) return createResponse({ success: false, message: "Failed", data: {} })
 
-  const typeUser = user as User
+  const typeUser = user as UserType
 
   const hashPassword: string = await encrypt(password)
   const response = await collection.updateOne({ email }, {$set: {password: hashPassword}})
@@ -185,4 +105,10 @@ export const changePassword = async(email: string, password: string) => {
     return createResponse({ success: true, message: "Signed in", data: serializedUserInfo })
   }
   return createResponse({ success: false, message: "Failed", data: {} })
+}
+
+export const checkUser = async (email: string) => {
+  const { user, collection } = await fetchUser(email)
+  if(user) return true
+  return false
 }
