+++ src/middleware.ts
@@ -4,7 +4,7 @@ import type { NextRequest } from 'next/server'
  
 // This function can be marked `async` if using `await` inside
 export function middleware(request: NextRequest) {
-  // console.log(request.cookies)
+  
   const pathName: string = request.nextUrl.pathname
   let redirectStatus = false
   if(pathName === "/signup" || pathName === "/signin") redirectStatus = true
