+++ src/types/Form.ts
@@ -1,0 +1,8 @@
+export type FormStatusType = "signin" | "signup" | "forgetPassword"
+
+export type SignUpFormDataType = {
+  name: string,
+  email: string,
+  password: string,
+  rePassword: string
+}
